@@ -75,31 +75,31 @@
     "mongoose": "^6.3.3",
     "morgan": "^1.10.0",
     "polished": "^4.2.2",
-    "prettier": "^2.6.2",
+    "prettier": "2.7.1",
     "pretty-quick": "^3.1.3",
     "qrcode": "^1.5.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-helmet": "^6.1.0",
     "react-hook-form": "^7.31.0",
     "react-i18next": "^11.16.9",
-    "react-native": "0.64.3",
+    "react-native": "0.70.1",
     "react-native-async-storage": "^0.0.1",
     "react-native-clock-analog": "^0.0.3",
     "react-native-date-picker": "^4.2.5",
-    "react-native-gesture-handler": "2.1.0",
+    "react-native-gesture-handler": "2.5.0",
     "react-native-get-random-values": "^1.8.0",
     "react-native-qrcode-svg": "^6.1.2",
-    "react-native-reanimated": "2.3.1",
-    "react-native-safe-area-context": "3.3.2",
-    "react-native-screens": "3.10.2",
-    "react-native-svg": "12.1.1",
+    "react-native-reanimated": "2.9.1",
+    "react-native-safe-area-context": "4.3.1",
+    "react-native-screens": "3.15.0",
+    "react-native-svg": "13.2.0",
     "react-native-svg-transformer": "1.0.0",
     "react-native-toast-notifications": "^3.2.3",
-    "react-native-web": "0.17.5",
+    "react-native-web": "0.18.10",
     "react-portal": "^4.2.2",
-    "react-redux": "^8.0.1",
-    "react-router-dom": "^5.3.0",
+    "react-redux": "8.0.2",
+    "react-router-dom": "6.4.1",
     "react-router-native": "5.2.1",
     "react-toastify": "^9.0.1",
     "regenerator-runtime": "0.13.7",
@@ -108,7 +108,7 @@
     "styled-components": "^5.3.5",
     "styled-container-query": "^1.3.5",
     "tslib": "^2.3.0",
-    "typescript": "^4.6.4",
+    "typescript": "4.8.4",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
