@@ -4,7 +4,7 @@ import { NotAuthorizedLayout } from "apps/web/src/components/NotAuthorizedLayout
 import { LinkService } from "apps/web/src/services/LinkService";
 import { Helmet } from "react-helmet";
 import { useTranslation } from "react-i18next";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import { useCoach } from "../../../useCoach";
 import { RegisterForm as UncontrolledRegisterForm } from "./RegisterForm";
@@ -17,13 +17,13 @@ export const RegisterForm: React.FC = () => {
     operations: { register },
   } = useCoach();
 
-  const history = useHistory();
+  const navigate = useNavigate();
 
   useEffect(() => {
     if (isLogged) {
-      return history.push(LinkService.home());
+      return navigate(LinkService.home());
     }
-  }, [isLogged, history]);
+  }, [isLogged, navigate]);
 
   return (
     <NotAuthorizedLayout>
