@@ -1,10 +1,10 @@
 import { DateInfo } from "@routine-support/domains";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import { updateCurrentDateInfoQuery } from "../utils/updateCurrentDateInfoQuery";
 
 export const useUpdateCurrentDateInfoQuery = () => {
-  const history = useHistory();
+  const navigate = useNavigate();
 
-  return (date: DateInfo) => updateCurrentDateInfoQuery(date, history);
+  return (date: DateInfo) => updateCurrentDateInfoQuery(date, navigate);
 };
