@@ -1,12 +1,14 @@
-import * as ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import "react-toastify/dist/ReactToastify.css";
 import App from "./app/app";
 import { AppWrapper } from "./components/AppWrapper";
 
-ReactDOM.render(
+const domContainer = document.getElementById("root");
+const root = createRoot(domContainer as any);
+
+root.render(
   <AppWrapper>
     <App />
-  </AppWrapper>,
-  document.getElementById("root")
+  </AppWrapper>
 );
